Rename directory and update imports
