+++ griptape/memory/task/storage/base_artifact_storage.py
